fixed add toString()

+++ app/src/main/java/com/youoweme/model/FixedPointDouble.kt
@@ -13,7 +13,7 @@ value class FixedPointDouble(val value: Int) : Comparable<FixedPointDouble> {
     override fun compareTo(other: FixedPointDouble) = value.compareTo(other.value)
 
     override fun toString(): String {
-        return value.toString()
+        return value.toDouble().toString()
     }
 
     /*
